requirement.txt
PyQt6>=6.4.0
pandas>=1.3.0
numpy>=1.21.0
pycryptodomex
openpyxl>=3.0.0
psutil>=5.9.0
xlrd